% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_log}
\alias{add_log}
\alias{debug_log}
\alias{error_log}
\alias{info_log}
\alias{trace_log}
\alias{warn_log}
\title{Sends a message to the currently active logging outputs}
\usage{
add_log(x, level, ...)

error_log(x, ...)

warn_log(x, ...)

info_log(x, ...)

debug_log(x, ...)

trace_log(x, ...)
}
\arguments{
\item{x}{an object to output in the log, such as a character vector}

\item{level}{the detail level of the message}

\item{...}{optional arguments passed to the low-level functions that
construct log messages; main usage is to exert control over arguments to
\code{print.data.frame}, such as \code{right} for alignment and
\code{row.names}. (See \code{format_message} for details.)}
}
\description{
Sends a message to the currently active logging outputs
}
\section{Functions}{
\itemize{
\item \code{}: Sends a message at the "ERROR" level

\item \code{}: Sends a message at the "WARN" level

\item \code{}: Sends a message at the "INFO" level

\item \code{}: Sends a message at the "DEBUG" level

\item \code{}: Sends a message at the "TRACE" level
}}

