% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logr.R
\name{log_error}
\alias{log_error}
\title{Logs an error}
\usage{
log_error(msg = NULL)
}
\arguments{
\item{msg}{The message to log.}
}
\description{
Writes an error message to the log. Error will be written
both to the log at the point the function is called, and also written to the
message file.  This function is used internally.
}
\examples{
library(logr)

# Create temp file location
tmp <- file.path(tempdir(), "test.log")

# Open log
lf <- log_open(tmp)

# Send error message to log
log_error("Here is a error")

# Close log
log_close()

# View results
writeLines(readLines(lf))

}
