% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplements.R
\name{log_path}
\alias{log_path}
\title{Get the path of the current log}
\usage{
log_path()
}
\value{
The full path to the currently opened log, or NULL if no log is open.
}
\description{
The \code{log_path} function gets the path to the currently
opened log.  This function may be useful when you want to manipulate
the log in some way, and need the path.  The function takes no parameters.
}
\examples{
# Create temp file location
tmp <- file.path(tempdir(), "test.log")

# Open log
log_open(tmp)

# Get path
lf <- log_path()

# Close log
log_close()

lf
}
