% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{logistic}
\alias{logistic}
\title{plogis}
\usage{
logistic(x)
}
\arguments{
\item{x}{value at which the logisitc functio is to be evaluated}
}
\value{
density of logistic function at point x
}
\description{
plogis
}
\keyword{Internal}
