% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browser.R
\name{logR_browser}
\alias{logR_browser}
\title{Browse logR data}
\usage{
logR_browser()
}
\description{
Shiny application which simply displays the logR logs.
}
\examples{
library(data.table)
N <- 1e5
df <- data.frame(a = rnorm(N), b = sample(seq_len(as.integer(log(N))),N,TRUE))
dt <- as.data.table(df)
dfr <- logR(with(df, aggregate(a, list(b), sum)), in_rows=nrow(df))
dtr <- logR(dt[,.(a=sum(a)),,b], in_rows=nrow(dt))
err <- logR(sum(1,"a"))
war <- logR(cor(c(1,1),c(2,3)))
if(interactive()) logR_browser()
file.remove("LOGR.csv")
}
\seealso{
\link{logR}, \link{logR_query}
}

