% Generated by roxygen2: do not edit by hand
\name{logR_connect}
\alias{logR_connect}
\title{Connect logR to postgres}
\usage{
logR_connect(host = Sys.getenv("POSTGRES_HOST", "127.0.0.1"),
  port = Sys.getenv("POSTGRES_PORT", "5432"),
  dbname = Sys.getenv("POSTGRES_DB", "postgres"),
  user = Sys.getenv("POSTGRES_USER", "postgres"),
  password = Sys.getenv("POSTGRES_PASSWORD", "postgres"), quoted = FALSE)
}
\arguments{
\item{host}{character hostname/ip by default from ENV var `POSTGRES_HOST`}

\item{port}{character port by default from ENV var `POSTGRES_PORT`}

\item{dbname}{character port by default from ENV var `POSTGRES_DB`}

\item{user}{character port by default from ENV var `POSTGRES_USER`}

\item{password}{character port by default from ENV var `POSTGRES_PASSWORD`}

\item{quoted}{logical if TRUE it will return unevaluated expression which can be passed to remote R session. Database authentication will be included in expression.}
}
\value{
logical, if successfully connected then TRUE, otherwise FALSE. Use \code{getOption("logR.conn")} to access actual connection, which is set as side effect.
}
\description{
Wrapper on \code{DBI::dbConnect} with defaults from envir.
}

