% Generated by roxygen2: do not edit by hand
\name{logR_disconnect}
\alias{logR_disconnect}
\title{Disconnect logR from postgres}
\usage{
logR_disconnect(quoted = FALSE)
}
\arguments{
\item{quoted}{logical if TRUE it will return unevaluated expression which can be passed to remote R session. Database authentication will be included in expression.}
}
\description{
Simple wrapper to \code{DBI::dbDisconnect}.
}

