% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logR.R
\name{logR_schema}
\alias{logR_schema}
\title{Populate logR schema}
\usage{
logR_schema(conn.name = getOption("logR.conn"), db_vendor = "H2",
  drop = FALSE)
}
\arguments{
\item{conn.name}{character name of defined db connection.}

\item{db_vendor}{character, currently supported \emph{H2} and \emph{MSSQL}.}

\item{drop}{logical, should be objects dropped if exists.}
}
\description{
Populate logR schema
}

