% Generated by roxygen2: do not edit by hand
\name{logR_watcher}
\alias{logR_watcher}
\title{Detects invalid logs}
\usage{
logR_watcher(since = Sys.time() - 86400)
}
\arguments{
\item{since}{POSIXt or Date, must be provided for this function.}
}
\value{
Will raise error in case if it will find any NULL stats logs. Otherwise NULL invisibly.
}
\description{
On unlikely fatal error R session crash while evaluting logged expression it will not update \code{status} field in database. This functions helps to detect such cases.
}
\details{
\code{since} parameter default value is current timestamp minus one day, so it should be scheduled daily. If you don't want to provide \code{since} param then use \code{logR_query} directly.
}
\seealso{
\link{logR_query}
}

