% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{neg_loglik}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{parameter vector}

\item{design}{design matrix for the logistic regression}

\item{response}{response vector}
}
\value{
the negative log-likelhood of the passed parameter given the passed
data
}
\description{
calculates the derivative of the negative log-likelihood of the
logistic model
}
\keyword{Internal}
