% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_sql}
\alias{schema_sql}
\title{logR schema scripts dictionary}
\usage{
schema_sql(table = getOption("logR.table"),
  seq_view = getOption("logR.seq_view"))
}
\arguments{
\item{table}{character table name for log storing.}

\item{seq_view}{character name of view which will query sequence.}
}
\description{
logR schema scripts dictionary
}
\note{
You can PR new database scripts.
}
\examples{
library(data.table)
# scripts for each vendor
schema_sql()[, .(sql = names(.SD)), vendor]
# create view statements for each vendor
schema_sql()[, create_view, vendor]
# print scripts for one vendor
invisible(schema_sql()["sqlserver", cat(paste(.SD,collapse="\\n")), .SDcols=-"vendor"])
}
\seealso{
\link{logR_schema}, \link{logR}
}

