% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logR.R
\name{sql_val}
\alias{sql_val}
\title{Prepare SQL values}
\usage{
sql_val(col, x, trunc_char_n = 252L)
}
\arguments{
\item{col}{character vector of column names.}

\item{x}{data.table a one row data.table.}

\item{trunc_char_n}{integer number of characters to truncate character after. Default 252, \emph{...} is added to the end of string so the result will have 255 characters.}
}
\value{
character vector
}
\description{
Wraps character types into single quotes, numbers as non scientific.
}

