% Generated by roxygen2: do not edit by hand
\name{tryCatch2}
\alias{tryCatch2}
\title{tryCatch2 for all conditions}
\usage{
tryCatch2(expr)
}
\arguments{
\item{expr}{expression to evaluate with exception handling.}
}
\value{
a list with elements 'value', 'error', 'warning', 'message', 'interrupt'.
}
\description{
Catch all conditions and keep computed results.
}

