% Generated by roxygen2: do not edit by hand
\name{tryCatch_we}
\alias{tryCatch_we}
\title{tryCatch both warnings and errors}
\usage{
tryCatch_we(expr)
}
\arguments{
\item{expr}{expression to evaluate with warnings and error handling.}
}
\value{
a list with 'value' and 'warning', where 'value' may be an error caught.
}
\description{
We want to catch *and* save both errors and warnings, and in the case of a warning, also keep the computed result.
}
\note{
Modified version to catch multiple warnings.
}
\author{
Martin Maechler
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2010-December/262626.html}
}

