% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{logr}
\alias{logr}
\alias{logr-package}
\title{Logging framework in the R idiom}
\description{
This package implements a logging framework that uses R's existing
messaging functionality, and builds upon it an API that is simple
to use. Because it shares the same underpinnings, \emph{logr} can
capture messages and warnings generated by \code{warning} and
\code{message}, making it easy to adopt even in a mature codebase.
}
\details{
\emph{Logr} puts these goals over likeness to Log4j, the logging
package popular in the Java world, even if it means sacrificing
some of Log4j's capabilities (and, I hope, the baroqueness of its
API) in the process.
}

