% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{logR}
\alias{logR}
\alias{logR-package}
\title{logR}
\usage{
logR(expr, gcFirst = FALSE, silent = FALSE, mail = list(),
  tag = NA_character_, in_rows = NA_integer_, conn = NULL,
  log_table = c("public", "logR"), verbose = 0, ddl = FALSE,
  ddl.purge = FALSE)
}
\arguments{
\item{expr}{expression to measure and log}

\item{gcFirst}{clean cache, use when comparing timings}

\item{silent}{do not raise warning/error, no email alert too}

\item{mail}{a list of args to be passed to \code{\link{send.mail}}, body element will be overwritten by alert.}

\item{tag}{character}

\item{in_rows}{integer input DT nrow}

\item{conn}{database connection}

\item{log_table}{length 2 vector of schema and table name, default \code{c("public","logR")}}

\item{verbose}{integer print status messages to console}

\item{ddl}{only create table for logR logs, skip processing}

\item{ddl.purge}{logical if drop exists log table and recreate}
}
\value{
result of expr, unless it's error
}
\description{
logR

Complete log solution, log status, timing, in-out rows to database. In case of alert also send email.
}

