% Generated by roxygen2: do not edit by hand
\name{logR}
\alias{logR}
\title{Detailed logging of R expressions}
\usage{
logR(expr, alert = TRUE, in_rows = NA_integer_,
  meta = getOption("logR.meta"), silent = getOption("logR.silent"),
  mail = getOption("logR.mail"), mail_args = getOption("logR.mail_args"),
  .conn = getOption("logR.conn"), .schema = getOption("logR.schema"),
  .table = getOption("logR.table"), .log = getOption("logR.log"))
}
\arguments{
\item{expr}{expression to be evaluted with logging.}

\item{alert}{logical, should be alert flag suppressed on warning/error, including suppressing email notification.}

\item{in_rows}{integer input DF/DT nrow, \emph{logR} can only guess \emph{out_rows}.}

\item{meta}{list, list of custom fields, each of length 1, list be always the same, fill missing with NA. Default \code{getOption("logR.meta",list())} means no meta columns. If you want to change element you need to alter table before.}

\item{silent}{logical, if default \code{getOption("logR.silent",TRUE)} it will not raise warning or error but only log.}

\item{mail}{logical if \emph{TRUE} then on alert the email will be send. Requires \emph{mail_args} to be provided. Default \code{getOption("logR.mail",FALSE)}.}

\item{mail_args}{list - mail not implemented yet.}

\item{.conn}{DBI connection. Default to \code{getOption("logR.conn",NULL)}.}

\item{.schema}{character scalar, location in database to store logs, default \code{getOption("logR.schema")}.}

\item{.table}{character scalar, location in database to store logs, default \code{getOption("logR.table")}.}

\item{.log}{logical escape parameter, set to \emph{FALSE} to suppress logR process and just execute a call, default to \code{getOption("logR.log",TRUE)}.}
}
\value{
Result of evaluated \emph{expr}.
}
\description{
Complete logging solution. Writes to database call metadata before its evaluation. Evalutes with timing and warning/error catching. Updates database entry for processing details: in/out rows, custom metadata, warning/error messages. Email on warning/errors.
}
\note{
Only first warning/message will be logged to database and send on email.
}
\section{Side effects}{

\itemize{
\item entry in database table.
\item in case of alerts and email notification configured also the email will be send.
}
}

\section{Database setup}{

You can create all three objects automatically using \link{logR_schema} function.
}

\section{Fatal errors}{

If your R function will manage to kill whole R session you will see that \emph{status} entry in log table will not get updated and it will stay as \emph{NA}.
It might be worth to schedule a watcher task to detect such cases, see \link{logR_watcher} vignette.
}
\seealso{
\link{logR_schema}, \link{logR_query}
}

