% Generated by roxygen2: do not edit by hand
\name{logR_query}
\alias{logR_query}
\title{Query logR data}
\usage{
logR_query(alert, status, since, .conn = getOption("logR.conn"),
  .table = getOption("logR.table"), .schema = getOption("logR.schema"))
}
\arguments{
\item{alert}{logical, if TRUE then filter to alerts.}

\item{status}{logical, if TRUE then filter to non-success.}

\item{since}{POSIXt or Date, if provided then filter on start timestamp.}

\item{.conn}{DBI connection.}

\item{.table}{character.}

\item{.schema}{character.}
}
\description{
Query logR logs from defined database or csv file.
}
\details{
By default all function arguments will be taken from options which are used to setup logR logging, for arguments description see \link{logR}.
}
\seealso{
\link{logR}
}

