% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browser.R
\name{logR_query}
\alias{logR_query}
\title{Query logR data}
\usage{
logR_query(.db = getOption("logR.db"), .conn = getOption("logR.conn"),
  .table = getOption("logR.table"),
  .db.conns = getOption("dwtools.db.conns"), .wd = getOption("logR.wd",
  getwd()))
}
\arguments{
\item{.db}{logical.}

\item{.conn}{character.}

\item{.table}{character.}

\item{.db.conns}{list of connection defined for \link[dwtools]{db}.}

\item{.wd}{character working directory path to be used when logging to csv file.}
}
\description{
Query logR logs from defined database or csv file.
}
\details{
By default all function arguments will be taken from options which are used to setup logR logging, for arguments description see \link{logR}.
}
\examples{
logR(Sys.sleep(0.01))
logR_query()
}
\seealso{
\link{logR}, \link{logR_browser}
}

