% Generated by roxygen2: do not edit by hand
\name{logR_schema}
\alias{logR_schema}
\title{Populate logR schema}
\usage{
logR_schema(meta = list(), drop = FALSE, silent = TRUE,
  .conn = getOption("logR.conn"))
}
\arguments{
\item{meta}{list of metadata columns, postgres data types as character string named by column name like \code{list(tag='VARCHAR(255)', val_col='DOUBLE PRECISION', int_col='INTEGER', systimecol='TIMESTAMPTZ')}. Avoid reserved postgres keywords.}

\item{drop}{logical, drop before creation.}

\item{silent}{logical, silently try calls to db}

\item{.conn}{DBI connection.}
}
\description{
Function will execute scripts create schema and table to store logs. To view scripts see \link{schema_sql}.
}
\seealso{
\link{schema_sql}, \link{logR}
}

