% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schema.R
\name{logR_schema}
\alias{logR_schema}
\title{Populate logR schema}
\usage{
logR_schema(conn.name = getOption("logR.conn"), vendor = c("h2",
  "sqlserver", "postgres", "oracle"), drop = FALSE)
}
\arguments{
\item{conn.name}{character name of defined db connection. See examples.}

\item{vendor}{character, currently supported \code{c("h2","sqlserver","postgres","oracle")}.}

\item{drop}{logical, try drop before creation.}
}
\description{
Populate logR schema
}
\examples{
# librarty()
# setup connection

#logR_schema()

#library(dwtools)
#db("SELECT * FROM DUAL;")
}
\seealso{
\link{schema_sql}
}

