% Generated by roxygen2: do not edit by hand
\name{logR_schema}
\alias{logR_schema}
\title{Populate logR schema}
\usage{
logR_schema(meta = list(), .conn = getOption("logR.conn"), drop = FALSE)
}
\arguments{
\item{.conn}{DBI connection.}

\item{drop}{logical, try drop before creation.}
}
\description{
There are three database objects required, plus optional schema, all are populated by this function call. To view scripts see \link{schema_sql}.
}
\seealso{
\link{schema_sql}, \link{logR}
}

