% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schema.R
\name{logR_schema}
\alias{logR_schema}
\title{Populate logR schema}
\usage{
logR_schema(conn.name = getOption("logR.conn"), vendor = c("h2",
  "sqlserver", "postgres", "oracle"), drop = FALSE)
}
\arguments{
\item{conn.name}{character name of defined db connection. See examples.}

\item{vendor}{character, currently supported \code{c("h2","sqlserver","postgres","oracle")}.}

\item{drop}{logical, try drop before creation.}
}
\description{
There are three database objects required, all are populated by this function call. To view scripts see \link{schema_sql}.
}
\examples{
if(requireNamespace("RH2",quietly=TRUE)){
  library(RH2)
  # define connection
  h2 <- list(drvName = "JDBC", conn = dbConnect(H2(), "jdbc:h2:mem:"))
  # setup options and connection
  opts <- options("dwtools.db.conns"=list(h2=h2),
                  "logR.db" = TRUE,
                  "logR.conn" = "h2")
  # run build schema scripts
  logR_schema(vendor = "h2")

  # check if exists
  library(dwtools)
  db("SELECT sequence_name, current_value FROM INFORMATION_SCHEMA.SEQUENCES")
  db("SELECT table_type, table_name FROM INFORMATION_SCHEMA.TABLES WHERE table_schema != 'INFORMATION_SCHEMA'")
}
}
\seealso{
\link{schema_sql}
}

