% Generated by roxygen2: do not edit by hand
\name{schema_sql}
\alias{schema_sql}
\title{logR schema scripts dictionary}
\usage{
schema_sql(schema = getOption("logR.schema"),
  table = getOption("logR.table"), meta = list())
}
\arguments{
\item{schema}{character schema name for log storing.}

\item{table}{character table name for log storing.}

\item{meta}{list of metadata columns, postgres data types as character string named by column name like \code{list(tag='VARCHAR(255)', val_col='DOUBLE PRECISION', int_col='INTEGER', systimecol='TIMESTAMPTZ')}. Avoid reserved postgres keywords.}
}
\description{
Use \link{logR_schema} to execute all scripts at once.
}
\examples{
sql = schema_sql(schema="logr")
invisible(lapply(sql, sapply, cat, sep="\\n"))
}
\seealso{
\link{logR_schema}, \link{logR}
}

