% Generated by roxygen2: do not edit by hand
\name{schema_sql}
\alias{schema_sql}
\title{logR schema scripts dictionary}
\usage{
schema_sql(table = getOption("logR.table"),
  seq_view = getOption("logR.seq_view"), schema = getOption("logR.schema"),
  meta = list())
}
\arguments{
\item{table}{character table name for log storing.}

\item{seq_view}{character name of view which will query sequence.}

\item{schema}{character schema name for log storing.}

\item{meta}{list of metadata columns, postgres data types as character string named by column name. Avoid reserved postgres keyword.}
}
\description{
logR schema scripts dictionary
}
\seealso{
\link{logR_schema}, \link{logR}
}

