% Generated by roxygen2: do not edit by hand
\name{update_make_set}
\alias{update_make_set}
\title{Make SET for update statement}
\usage{
update_make_set(col, x)
}
\arguments{
\item{col}{character vector of column names to update.}

\item{x}{data.table a one row data.table.}
}
\value{
character vector which can be easily \code{paste(x,collapse=",")} to paste in update statement.
}
\description{
Helper for producing update statement, wraps character types into single quotes, numbers as non scientific.
}

